De-duplicate the assemble function
Whoops, forgot I had a whole 'assemble' module
